refactor: rename routers to RawEncoded..Router
